# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a small api to help you clean up hieradata",
        "contact": {},
        "license": {},
        "version": "0.0.3"
    },
    "basePath": "/v1",
    "paths": {
...
        "/hiera/path": {
            "get": {
                "description": "Gets all the ids of your paths so you can see which hiera paths are available.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get all hiera path ids",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.APIArrayMessage"
                        }
                    },
                    "500": {
                        "description": "Something went wrong getting the entry",
                        "schema": {
                            "$ref": "#/definitions/api.APIMessage"
                        }
                    }
                }
            }
        },
        "/hiera/path/{id}": {
            "get": {
                "description": "Get the data from one hiera path",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get a hiera path",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Some ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Something went wrong getting the entry",
                        "schema": {
                            "$ref": "#/definitions/api.APIMessage"
                        }
                    }
                }
            },
            "put": {
                "description": "Creates a new hiera path entry if it does not exist yet.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Updates an existing hiera path entry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Some ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.HieraDataExample"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.APIMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.APIMessage"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new hiera path entry if it does not exist yet.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Creates a hiera path entry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Some ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.HieraDataExample"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.APIMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.APIMessage"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a hiera path entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a hiera path",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Some ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.APIMessage"
                        }
                    },
                    "500": {
                        "description": "Something went wrong getting the entry",
                        "schema": {
                            "$ref": "#/definitions/api.APIMessage"
                        }
                    }
                }
            }
        },
...
        "/hiera/variable/hierarchy": {
            "get": {
                "description": "Gets all the variable paths that are in your configuration file.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get all variable paths in your configuration",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.APIArrayMessage"
                        }
                    },
                    "500": {
                        "description": "Something went wrong getting the entry",
                        "schema": {
                            "$ref": "#/definitions/api.APIMessage"
                        }
                    }
                }
            }
        },
        "/hiera/variable/hierarchy/{id}": {
            "get": {
                "description": "Translates the",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get all variable paths for a specific host",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Some ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.HierarchyResult"
                        }
                    },
                    "500": {
                        "description": "Something went wrong getting the entry",
                        "schema": {
                            "$ref": "#/definitions/api.APIMessage"
                        }
                    }
                }
            }
        },
...
        "/hierarchy/{id}": {
            "get": {
                "description": "Transaltes the hierarchies in your hiera file into actual paths. By getting the facts from puppetdb.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get the hierachies for a specific host.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Some ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.HierarchyResult"
                        }
                    },
                    "500": {
                        "description": "Something went wrong getting the entry",
                        "schema": {
                            "$ref": "#/definitions/api.APIMessage"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.APIArrayMessage": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "api.APIMessage": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "api.HieraDataExample": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "key2": {
                    "type": "object",
                    "additionalProperties": true
                },
                "key3": {
                    "type": "boolean"
                },
                "key4": {
                    "type": "integer"
                }
            }
        },
...
        "api.HierarchyResult": {
            "type": "object",
            "properties": {
                "paths": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "vars": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    }
}